Moviendo `randomString` de `Controller` a `SecurityTools`
Quién sabe qué hacía esa función en Controller. Este cambio pone esa
función en un lugar más sensato.
