+++ frontend/server/controllers/InterviewController.php
@@ -106,7 +106,7 @@ class InterviewController extends Controller {
             $newUserRequest = new Request($r);
             $newUserRequest['email'] = $r['usernameOrEmail'];
             $newUserRequest['username'] = UserController::makeUsernameFromEmail($r['usernameOrEmail']);
-            $newUserRequest['password'] = self::randomString(8);
+            $newUserRequest['password'] = SecurityTools::randomString(8);
             $newUserRequest['skip_verification_email'] = 1;
 
             UserController::apiCreate($newUserRequest);
